App to MySQL connection done

+++ src/main/java/com/example/spring_boot_practice/repository/TodoRepository.java
@@ -4,7 +4,6 @@ import com.example.spring_boot_practice.model.Todo;
 import org.springframework.stereotype.Repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-@Repository // Repositoryクラスであることを示す
-public interface TodoRepository extends JpaRepository<Todo, Long> { 
-
+@Repository
+public interface TodoRepository extends JpaRepository<Todo, Integer> {
 }
